* master:
  Removed unused method in ContentManager
  - Correct handling of the server user token expiration
  Removed unnecessary allocation of memory when creating tags, turned out values did not get overwritten, even though it looked like it when debugging. Thank you Erik for discovering this :).
  Reverted to using the 3rd party lib LoopVIewPager as per Eriks request.
  Removed 3rd party lib reference to LoopViewPager, and using an own modified version of it, which should fix the ArithmeticException: divide by zero causing crash
  Trying to fix NPE bug causing crash.
  All methods in Cache and in PersistentCache that was not synchronised, are now synchronised. Changed use of cache.someMethod in ContentManager to getCache().someMethod, where getCache() reinitialises the Cache if it is null, trying to fix NPE bug where Cache is null.
  Comment
  Changed Analytics for event Sign up, to separate between Facebook and Email signup. Also changed "goal" on Google Analytics Dashboard to match BEGINS WITH rather than equals "SIGN_UP_COMPLETED"
  Fixing ArithmeticException: divide by zero caused by the LoopViewPager 3rd party lib
  Documentation of the data fetching flow
  - ContentManager is now part of the SecondScreenApplication class - Fix for Facebook login flow (if Facebook app not installed)
  Added a build and a properties file for jenkins to use

Conflicts:
	res/values/strings.xml

+++ src/com/mitv/PersistentCache.java
@@ -130,7 +130,7 @@ public abstract class PersistentCache
 	}
 	
 	
-	public boolean containsTVDates() 
+	public synchronized boolean containsTVDates() 
 	{
 		boolean containsTVDates = (tvDates != null && !tvDates.isEmpty());
 		
@@ -225,12 +225,15 @@ public abstract class PersistentCache
 	}
 	
 	
-	public void clearUserData() 
+	public synchronized void clearUserData() 
 	{
-		UserLoginDataORM userLoginDataORM = new UserLoginDataORM(userData);
-		userLoginDataORM.delete();
+		if(userData != null)
+		{
+			UserLoginDataORM userLoginDataORM = new UserLoginDataORM(userData);
+			userLoginDataORM.delete();
 		
-		userData = null;
+			userData = null;
+		}
 	}
 	
 	
@@ -585,7 +588,7 @@ public abstract class PersistentCache
 	
 	/* TV CHANNEL IDS USER */
 	
-	public void setTvChannelIdsUser(ArrayList<TVChannelId> tvChannelIdsUser) 
+	public synchronized void setTvChannelIdsUser(ArrayList<TVChannelId> tvChannelIdsUser) 
 	{
 		this.tvChannelIdsUser = tvChannelIdsUser;
 		
@@ -596,7 +599,7 @@ public abstract class PersistentCache
 	
 	/* TV GUIDES MY */
 	
-	protected void clearTVGuidesMy() 
+	protected synchronized void clearTVGuidesMy() 
 	{
 		tvGuidesMy = null;
 	}
@@ -648,7 +651,7 @@ public abstract class PersistentCache
 	
 	/* TV GUIDES ALL */
 	
-	protected TVGuide getTVGuideUsingTVDateNonFiltered(TVDate tvDate) 
+	protected synchronized TVGuide getTVGuideUsingTVDateNonFiltered(TVDate tvDate) 
 	{
 		String tvDateId = tvDate.getId();
